--- app/dashboard/admin/manage-users/page.tsx
@@ -1,489 +1,0 @@
-"use client";
-
-import { useState } from "react";
-import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import { Badge } from "@/components/ui/badge";
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardHeader,
-  CardTitle,
-} from "@/components/ui/card";
-import {
-  Table,
-  TableBody,
-  TableCell,
-  TableHead,
-  TableHeader,
-  TableRow,
-} from "@/components/ui/table";
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "@/components/ui/select";
-import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-} from "@/components/ui/dialog";
-import { Label } from "@/components/ui/label";
-import {
-  Search,
-  MoreHorizontal,
-  UserPlus,
-  Ban,
-  CheckCircle,
-  Edit,
-  Trash2,
-} from "lucide-react";
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuLabel,
-  DropdownMenuSeparator,
-  DropdownMenuTrigger,
-} from "@/components/ui/dropdown-menu";
-import { UserRole } from "@/types/enums";
-import { useEffect } from "react";
-
-export default function ManageUsersPage() {
-  const [users, setUsers] = useState<any[]>([]);
-  const [isLoading, setIsLoading] = useState(true);
-  const [totalUsers, setTotalUsers] = useState(0);
-  const [searchTerm, setSearchTerm] = useState("");
-  const [roleFilter, setRoleFilter] = useState("all");
-  const [statusFilter, setStatusFilter] = useState("all");
-  const [selectedUser, setSelectedUser] = useState<any>(null);
-  const [isEditDialogOpen, setIsEditDialogOpen] = useState(false);
-  const [isBanDialogOpen, setIsBanDialogOpen] = useState(false);
-
-  useEffect(() => {
-    const fetchUsers = async () => {
-      setIsLoading(true);
-      try {
-        const query = new URLSearchParams({
-          role: roleFilter !== "all" ? roleFilter : "",
-          page: "1",
-          limit: "100",
-        }).toString();
-
-        const res = await fetch(`/api/users?${query}`);
-        const data = await res.json();
-
-        const formattedUsers = data.users.map((user: any) => {
-          return {
-            ...user,
-            id: user._id,
-            status: user.status || "active",
-            joinedAt: new Date(user.createdAt).toLocaleDateString(),
-          };
-        });
-
-        setUsers(formattedUsers);
-        setTotalUsers(data.total);
-      } catch (err) {
-        console.error("Failed to fetch users", err);
-      } finally {
-        setIsLoading(false);
-      }
-    };
-
-    fetchUsers();
-  }, [roleFilter]);
-
-  const filteredUsers = users.filter((user) => {
-    const matchesSearch =
-      user.name.toLowerCase().includes(searchTerm.toLowerCase()) ||
-      user.email.toLowerCase().includes(searchTerm.toLowerCase());
-    const matchesRole = roleFilter === "all" || user.role === roleFilter;
-    const matchesStatus =
-      statusFilter === "all" || user.status === statusFilter;
-    return matchesSearch && matchesRole && matchesStatus;
-  });
-
-  const handleBanUser = (user: any) => {
-    setSelectedUser(user);
-    setIsBanDialogOpen(true);
-  };
-
-  const handleEditUser = (user: any) => {
-    setSelectedUser(user);
-    setIsEditDialogOpen(true);
-  };
-  const confirmBan = async () => {
-    if (!selectedUser) return;
-
-    const newStatus = selectedUser.status === "banned" ? "active" : "banned";
-
-    try {
-      const res = await fetch(`/api/users/${selectedUser.id}/status`, {
-        method: "PATCH",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify({ status: newStatus }),
-      });
-
-      if (!res.ok) throw new Error("Failed to update user status");
-
-      // Cập nhật lại state sau khi server đã lưu
-      setUsers(
-        users.map((user) =>
-          user.id === selectedUser.id ? { ...user, status: newStatus } : user
-        )
-      );
-    } catch (err) {
-      console.error("Error updating status:", err);
-      alert("Failed to update user status.");
-    } finally {
-      setIsBanDialogOpen(false);
-      setSelectedUser(null);
-    }
-  };
-
-  const saveUserChanges = async () => {
-    if (!selectedUser) return;
-
-    try {
-      const res = await fetch(`/api/users/${selectedUser.id}`, {
-        method: "PATCH",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify({
-          name: selectedUser.name,
-          email: selectedUser.email,
-          role: selectedUser.role,
-          status: selectedUser.status,
-        }),
-      });
-
-      if (!res.ok) throw new Error("Failed to update user");
-
-      const { user: updatedUser } = await res.json();
-
-      setUsers((prev) =>
-        prev.map((user) =>
-          user.id === updatedUser._id
-            ? { ...updatedUser, id: updatedUser._id }
-            : user
-        )
-      );
-
-      setIsEditDialogOpen(false);
-      setSelectedUser(null);
-    } catch (err) {
-      console.error("Error saving user:", err);
-      alert("Failed to update user.");
-    }
-  };
-
-  const deleteUser = async (userId: string) => {
-    const confirm = window.confirm(
-      "Are you sure you want to delete this user?"
-    );
-    if (!confirm) return;
-
-    try {
-      const res = await fetch(`/api/users/${userId}`, {
-        method: "DELETE",
-      });
-
-      if (!res.ok) throw new Error("Delete failed");
-
-      setUsers((prev) => prev.filter((user) => user.id !== userId));
-    } catch (err) {
-      console.error("Delete user error:", err);
-      alert("Failed to delete user.");
-    }
-  };
-
-  const getStatusBadge = (status: string) => {
-    switch (status) {
-      case "active":
-        return (
-          <Badge variant="default" className="bg-green-100 text-green-800">
-            Active
-          </Badge>
-        );
-
-      case "banned":
-        return <Badge variant="destructive">Banned</Badge>;
-      default:
-        return <Badge variant="outline">{status}</Badge>;
-    }
-  };
-
-  const getRoleBadge = (role: string) => {
-    switch (role) {
-      case UserRole.ADMIN:
-        return (
-          <Badge variant="default" className="bg-purple-100 text-purple-800">
-            Admin
-          </Badge>
-        );
-      case UserRole.RECRUITER:
-        return (
-          <Badge variant="default" className="bg-blue-100 text-blue-800">
-            Recruiter
-          </Badge>
-        );
-      case UserRole.JOB_SEEKER:
-        return <Badge variant="outline">Job Seeker</Badge>;
-      default:
-        return <Badge variant="outline">{role}</Badge>;
-    }
-  };
-
-  return (
-    <div className="space-y-6">
-      <div className="flex justify-between items-center">
-        <div>
-          <h1 className="text-3xl font-bold">Manage Users</h1>
-          <p className="text-gray-600">View and manage all platform users</p>
-        </div>
-      </div>
-
-      <Card>
-        <CardHeader>
-          <CardTitle>User Management</CardTitle>
-          <CardDescription>
-            Total users: {totalUsers} | Displayed: {users.length}
-          </CardDescription>
-        </CardHeader>
-        <CardContent>
-          <div className="flex gap-4 mb-6">
-            <div className="relative flex-1">
-              <Search className="absolute left-3 top-3 h-4 w-4 text-gray-400" />
-              <Input
-                placeholder="Search users by name or email..."
-                value={searchTerm}
-                onChange={(e) => setSearchTerm(e.target.value)}
-                className="pl-10"
-              />
-            </div>
-            <Select value={roleFilter} onValueChange={setRoleFilter}>
-              <SelectTrigger className="w-40">
-                <SelectValue placeholder="Filter by role" />
-              </SelectTrigger>
-              <SelectContent>
-                <SelectItem value="all">All Roles</SelectItem>
-                <SelectItem value="admin">Admin</SelectItem>
-                <SelectItem value="recruiter">Recruiter</SelectItem>
-                <SelectItem value="job_seeker">Job Seeker</SelectItem>
-              </SelectContent>
-            </Select>
-            <Select value={statusFilter} onValueChange={setStatusFilter}>
-              <SelectTrigger className="w-40">
-                <SelectValue placeholder="Filter by status" />
-              </SelectTrigger>
-              <SelectContent>
-                <SelectItem value="all">All Status</SelectItem>
-                <SelectItem value="active">Active</SelectItem>
-                <SelectItem value="banned">Banned</SelectItem>
-              </SelectContent>
-            </Select>
-          </div>
-
-          <Table>
-            <TableHeader>
-              <TableRow>
-                <TableHead>User</TableHead>
-                <TableHead>Role</TableHead>
-                <TableHead>Status</TableHead>
-                <TableHead>Joined</TableHead>
-                <TableHead className="text-right">Actions</TableHead>
-              </TableRow>
-            </TableHeader>
-            <TableBody>
-              {filteredUsers.map((user) => (
-                <TableRow key={user.id}>
-                  <TableCell>
-                    <div className="flex items-center gap-3">
-                      <img
-                        src={user.avatar || "/placeholder.svg"}
-                        alt={user.name}
-                        className="w-8 h-8 rounded-full"
-                      />
-                      <div>
-                        <div className="font-medium">{user.name}</div>
-                        <div className="text-sm text-gray-500">
-                          {user.email}
-                        </div>
-                      </div>
-                    </div>
-                  </TableCell>
-                  <TableCell>{getRoleBadge(user.role)}</TableCell>
-                  <TableCell>{getStatusBadge(user.status)}</TableCell>
-                  <TableCell>{user.joinedAt}</TableCell>
-                  <TableCell className="text-right">
-                    <DropdownMenu>
-                      <DropdownMenuTrigger asChild>
-                        <Button variant="ghost" className="h-8 w-8 p-0">
-                          <MoreHorizontal className="h-4 w-4" />
-                        </Button>
-                      </DropdownMenuTrigger>
-                      <DropdownMenuContent align="end">
-                        <DropdownMenuLabel>Actions</DropdownMenuLabel>
-                        <DropdownMenuItem onClick={() => handleEditUser(user)}>
-                          <Edit className="mr-2 h-4 w-4" />
-                          Edit User
-                        </DropdownMenuItem>
-                        <DropdownMenuItem onClick={() => handleBanUser(user)}>
-                          {user.status === "banned" ? (
-                            <>
-                              <CheckCircle className="mr-2 h-4 w-4" />
-                              Unban User
-                            </>
-                          ) : (
-                            <>
-                              <Ban className="mr-2 h-4 w-4" />
-                              Ban User
-                            </>
-                          )}
-                        </DropdownMenuItem>
-                        <DropdownMenuSeparator />
-                        <DropdownMenuItem
-                          onClick={() => deleteUser(user.id)}
-                          className="text-red-600"
-                        >
-                          <Trash2 className="mr-2 h-4 w-4" />
-                          Delete User
-                        </DropdownMenuItem>
-                      </DropdownMenuContent>
-                    </DropdownMenu>
-                  </TableCell>
-                </TableRow>
-              ))}
-            </TableBody>
-          </Table>
-
-          {filteredUsers.length === 0 && (
-            <div className="text-center py-8 text-gray-500">
-              No users found matching your criteria.
-            </div>
-          )}
-        </CardContent>
-      </Card>
-
-      {/* Edit User Dialog */}
-      <Dialog open={isEditDialogOpen} onOpenChange={setIsEditDialogOpen}>
-        <DialogContent>
-          <DialogHeader>
-            <DialogTitle>Edit User</DialogTitle>
-            <DialogDescription>
-              Update user information and settings.
-            </DialogDescription>
-          </DialogHeader>
-          {selectedUser && (
-            <div className="space-y-4">
-              <div className="space-y-2">
-                <Label htmlFor="edit-name">Name</Label>
-                <Input
-                  id="edit-name"
-                  value={selectedUser.name}
-                  onChange={(e) =>
-                    setSelectedUser({ ...selectedUser, name: e.target.value })
-                  }
-                />
-              </div>
-              <div className="space-y-2">
-                <Label htmlFor="edit-email">Email</Label>
-                <Input
-                  id="edit-email"
-                  type="email"
-                  value={selectedUser.email}
-                  onChange={(e) =>
-                    setSelectedUser({ ...selectedUser, email: e.target.value })
-                  }
-                />
-              </div>
-              <div className="space-y-2">
-                <Label htmlFor="edit-role">Role</Label>
-                <Select
-                  value={selectedUser.role}
-                  onValueChange={(value) =>
-                    setSelectedUser({ ...selectedUser, role: value })
-                  }
-                >
-                  <SelectTrigger>
-                    <SelectValue />
-                  </SelectTrigger>
-                  <SelectContent>
-                    <SelectItem value="job_seeker">Job Seeker</SelectItem>
-                    <SelectItem value="recruiter">Recruiter</SelectItem>
-                    <SelectItem value="admin">Admin</SelectItem>
-                  </SelectContent>
-                </Select>
-              </div>
-              <div className="space-y-2">
-                <Label htmlFor="edit-status">Status</Label>
-                <Select
-                  value={selectedUser.status}
-                  onValueChange={(value) =>
-                    setSelectedUser({ ...selectedUser, status: value })
-                  }
-                >
-                  <SelectTrigger>
-                    <SelectValue />
-                  </SelectTrigger>
-                  <SelectContent>
-                    <SelectItem value="active">Active</SelectItem>
-                    <SelectItem value="banned">Banned</SelectItem>
-                  </SelectContent>
-                </Select>
-              </div>
-            </div>
-          )}
-          <DialogFooter>
-            <Button
-              variant="outline"
-              onClick={() => setIsEditDialogOpen(false)}
-            >
-              Cancel
-            </Button>
-            <Button onClick={saveUserChanges}>Save Changes</Button>
-          </DialogFooter>
-        </DialogContent>
-      </Dialog>
-
-      {/* Ban User Dialog */}
-      <Dialog open={isBanDialogOpen} onOpenChange={setIsBanDialogOpen}>
-        <DialogContent>
-          <DialogHeader>
-            <DialogTitle>
-              {selectedUser?.status === "banned" ? "Unban User" : "Ban User"}
-            </DialogTitle>
-            <DialogDescription>
-              {selectedUser?.status === "banned"
-                ? `Are you sure you want to unban ${selectedUser?.name}? They will regain access to the platform.`
-                : `Are you sure you want to ban ${selectedUser?.name}? They will lose access to the platform.`}
-            </DialogDescription>
-          </DialogHeader>
-          <DialogFooter>
-            <Button variant="outline" onClick={() => setIsBanDialogOpen(false)}>
-              Cancel
-            </Button>
-            <Button
-              variant={
-                selectedUser?.status === "banned" ? "default" : "destructive"
-              }
-              onClick={confirmBan}
-            >
-              {selectedUser?.status === "banned" ? "Unban User" : "Ban User"}
-            </Button>
-          </DialogFooter>
-        </DialogContent>
-      </Dialog>
-    </div>
-  );
-}
